agent/requirements.txt
google-adk>=1.21.0
python-dotenv>=1.0.0
google-cloud-aiplatform[adk,agent_engines]
google-cloud-storage